requirements.txt
PyYAML
websockets
cryptography
PyJWT
